students/k3340/Chernykh_Oleg/Lr1/fastapi-todo-app/requirements.txt
fastapi
pydantic
uvicorn
tortoise-orm
bcrypt
python-dotenv
python-jose
python-multipart
